+++ src/types.ts
@@ -1962,6 +1962,12 @@ export type ProjectErrors = {
   title?: Maybe<Scalars['String']['output']>;
 };
 
+/** Project search filter options */
+export type ProjectFilterOptions = {
+  /** Filter results by the plan's status */
+  status?: InputMaybe<PlanStatus>;
+};
+
 /** Funding that is supporting a research project */
 export type ProjectFunding = {
   __typename?: 'ProjectFunding';
@@ -2458,6 +2464,7 @@ export type QueryMetadataStandardsArgs = {
 
 
 export type QueryMyProjectsArgs = {
+  filterOptions?: InputMaybe<ProjectFilterOptions>;
   paginationOptions?: InputMaybe<PaginationOptions>;
   term?: InputMaybe<Scalars['String']['input']>;
 };
@@ -4177,6 +4184,7 @@ export type ResolversTypes = {
   ProjectCollaboratorAccessLevel: ProjectCollaboratorAccessLevel;
   ProjectCollaboratorErrors: ResolverTypeWrapper<ProjectCollaboratorErrors>;
   ProjectErrors: ResolverTypeWrapper<ProjectErrors>;
+  ProjectFilterOptions: ProjectFilterOptions;
   ProjectFunding: ResolverTypeWrapper<ProjectFunding>;
   ProjectFundingErrors: ResolverTypeWrapper<ProjectFundingErrors>;
   ProjectFundingStatus: ProjectFundingStatus;
@@ -4333,6 +4341,7 @@ export type ResolversParentTypes = {
   ProjectCollaborator: ProjectCollaborator;
   ProjectCollaboratorErrors: ProjectCollaboratorErrors;
   ProjectErrors: ProjectErrors;
+  ProjectFilterOptions: ProjectFilterOptions;
   ProjectFunding: ProjectFunding;
   ProjectFundingErrors: ProjectFundingErrors;
   ProjectImportInput: ProjectImportInput;